Rename directory to receive-message
