+++ app/routes/receive-message.js
@@ -7,27 +7,27 @@ const helpers = require('../../lib/helpers');
 const router = express.Router();
 bot.getBot();
 
-const paramsMiddleware = require('../../lib/middleware/chatbot/params');
+const paramsMiddleware = require('../../lib/middleware/receive-message/params');
 const getUserMiddleware = require('../../lib/middleware/user-get-by-platform');
 const createUserMiddleware = require('../../lib/middleware/user-create');
 
-const inboundMessageMiddleware = require('../../lib/middleware/chatbot/message-inbound');
+const inboundMessageMiddleware = require('../../lib/middleware/receive-message/message-inbound');
 
-const getBotReplyMiddleware = require('../../lib/middleware/chatbot/reply-brain');
-const brainTemplateMiddleware = require('../../lib/middleware/chatbot/template-brain');
-const noReplyMiddleware = require('../../lib/middleware/chatbot/template-paused');
+const getBotReplyMiddleware = require('../../lib/middleware/receive-message/reply-brain');
+const brainTemplateMiddleware = require('../../lib/middleware/receive-message/template-brain');
+const noReplyMiddleware = require('../../lib/middleware/receive-message/template-paused');
 const campaignMenuMiddleware = require('../../lib/middleware/campaign-menu');
 const campaignKeywordMiddleware = require('../../lib/middleware/campaign-keyword');
 const currentCampaignMiddleware = require('../../lib/middleware/campaign-current');
-const parseAskSignupMiddleware = require('../../lib/middleware/chatbot/parse-ask-signup-response');
-const parseAskContinueMiddleware = require('../../lib/middleware/chatbot/parse-ask-continue-response');
-const askContinueMiddleware = require('../../lib/middleware/chatbot/template-ask-continue');
+const parseAskSignupMiddleware = require('../../lib/middleware/receive-message/parse-ask-signup-response');
+const parseAskContinueMiddleware = require('../../lib/middleware/receive-message/parse-ask-continue-response');
+const askContinueMiddleware = require('../../lib/middleware/receive-message/template-ask-continue');
 const setUserCampaignMiddleware = require('../../lib/middleware/user-set-campaign');
-const gambitReplyMiddleware = require('../../lib/middleware/chatbot/template-gambit');
-const campaignMessageMiddleware = require('../../lib/middleware/chatbot/template-campaign');
+const gambitReplyMiddleware = require('../../lib/middleware/receive-message/template-gambit');
+const campaignMessageMiddleware = require('../../lib/middleware/receive-message/template-campaign');
 const setLastReplyTemplateMiddleware = require('../../lib/middleware/user-set-last-reply-template');
 
-const outboundMessageMiddleware = require('../../lib/middleware/chatbot/message-outbound');
+const outboundMessageMiddleware = require('../../lib/middleware/receive-message/message-outbound');
 
 router.use(paramsMiddleware());
