+++ lib/middleware/receive-message/message-inbound.js
