+++ lib/middleware/receive-message/message-outbound.js
