+++ lib/middleware/receive-message/parse-ask-continue-response.js
