+++ lib/middleware/receive-message/parse-ask-signup-response.js
