+++ lib/middleware/receive-message/reply-brain.js
