+++ lib/middleware/receive-message/template-ask-continue.js
