+++ lib/middleware/receive-message/template-brain.js
