+++ lib/middleware/receive-message/template-gambit.js
