+++ lib/middleware/receive-message/template-paused.js
